+++ mvcc/utils/LevelGenerator.go
@@ -6,12 +6,12 @@ import (
 )
 
 type LevelGenerator struct {
-	maxLevel   int
+	maxLevel   uint8
 	skipFactor int
 	random     *rand.Rand
 }
 
-func NewLevelGenerator(maxLevel int) LevelGenerator {
+func NewLevelGenerator(maxLevel uint8) LevelGenerator {
 	random := rand.New(rand.NewSource(time.Now().UnixNano()))
 
 	return LevelGenerator{
@@ -21,8 +21,8 @@ func NewLevelGenerator(maxLevel int) LevelGenerator {
 	}
 }
 
-func (levelGenerator LevelGenerator) Generate() int {
-	level := 1
+func (levelGenerator LevelGenerator) Generate() uint8 {
+	level := uint8(1)
 	newRandom := levelGenerator.random.Float64()
 	for level < levelGenerator.GetMaxLevel() && newRandom < 1.0/float64(levelGenerator.skipFactor) {
 		level = level + 1
@@ -31,6 +31,6 @@ func (levelGenerator LevelGenerator) Generate() int {
 	return level
 }
 
-func (levelGenerator LevelGenerator) GetMaxLevel() int {
+func (levelGenerator LevelGenerator) GetMaxLevel() uint8 {
 	return levelGenerator.maxLevel
 }
